# measurements/openapi/measurements.yml
swagger: '2.0'
info:
  title: OONI Measurements API
  version: '1.0.0'
  contact:
    name: Open Observatory of Network Interference (OONI)
    email: contact@openobservatory.org
    url: 'https://api.ooni.io/'
  x-logo:
    url: 'https://ooni.torproject.org/images/ooni-header-mascot-2x.png'
    backgroundColor: '#FFF'
basePath: '/api/v1'
paths:
  /measurements:
    get:
      summary: ''
      description: ''
      operationId: 'measurements.api.list_measurements'
      parameters:
        - name: report_id
          in: query
          type: string
          description: The report_id to search measurements for
        - name: input
          in: query
          type: string
          minLength: 3 # `input` is handled by pg_trgm
          description: The input (for example a URL or IP address) to search measurements for
        - name: domain
          in: query
          type: string
          minLength: 3
          description: The domain to search measurements for
        - name: probe_cc
          in: query
          type: string
          description: The two letter country code
        - name: probe_asn
          in: query
          type: string
          description: the Autonomous system number in the format "ASXXX"
        - name: test_name
          in: query
          type: string
          description: The name of the test
          enum:
          - web_connectivity
          - http_requests
          - dns_consistency
          - http_invalid_request_line
          - bridge_reachability
          - tcp_connect
          - http_header_field_manipulation
          - http_host
          - multi_protocol_traceroute
          - meek_fronted_requests_test
          - whatsapp
          - vanilla_tor
          - facebook_messenger
          - ndt
          - dash
          - telegram
        - name: since
          in: query
          type: string
          description: >-
            The start date of when measurements were run (ex.
            "2016-10-20T10:30:00")
        - name: until
          in: query
          type: string
          description: >-
            The end date of when measurement were run (ex.
            "2016-10-20T10:30:00")
        - name: since_index
          in: query
          type: string
          description: Return results only strictly greater than the provided index

        - name: confirmed
          in: query
          type: array
          collectionFormat: csv
          items:
            type: string
          description: |
            Will be true for confirmed network anomalies (we found a blockpage, a middlebox was found, the IM app is blocked, etc.).

        - name: anomaly
          in: query
          type: array
          collectionFormat: csv
          items:
            type: string
          description: |
            Measurements that require special attention (it's likely to be a case of blocking), however it has not nessarily been confirmed

        - name: failure
          in: query
          type: array
          collectionFormat: csv
          items:
            type: string
          description: |
            There was an error in the measurement (the control request failed, there was a bug, etc.).
            Default is to consider it both true or false (`failure=true,false`)

        - name: order_by
          in: query
          type: string
          description: 'By which key the results should be ordered by (default: `test_start_time`)'
          enum:
            - test_start_time
            - input
            - probe_cc
            - probe_asn
            - test_name
        - name: order
          in: query
          type: string
          description: |-
            If the order should be ascending or descending (one of: `asc` or `desc`)
          enum:
            - asc
            - desc
            - ASC
            - DESC
        - name: offset
          in: query
          type: integer
          description: 'Offset into the result set (default: 0)'
        - name: limit
          in: query
          type: integer
          description: 'Number of records to return (default: 100)'
      responses:
        '200':
          description: Returns the list of measurement IDs for the specified criteria
          schema:
            $ref: "#/definitions/MeasurementList"

        default:
          description: Default response

      x-code-samples:
        - lang: 'curl'
          source: |
            curl "https://api.ooni.io/api/v1/measurements?probe_cc=IT&confirmed=true&since=2017-09-01"

  /measurement/{measurement_id}:
    get:
      summary: ''
      description: ''
      operationId: 'measurements.api.get_measurement'
      parameters:
        - name: measurement_id
          in: path
          required: true
          type: string
          description: The measurement_id to retrieve the measurement for
        - name: download
          in: query
          type: boolean
          description: If we should be triggering a file download
      responses:
        '200':
          description: Returns the JSON blob for the specified measurement
          schema:
            $ref: "#/definitions/MeasurementBlob"

        default:
          description: Default response
  /files:
    get:
      summary: ''
      description: ''
      operationId: 'measurements.api.list_files'
      parameters:
        - name: probe_cc
          in: query
          type: string
          description: The two letter country code
        - name: probe_asn
          in: query
          type: string
          description: the Autonomous system number in the format "ASXXX"
        - name: test_name
          in: query
          type: string
          description: The name of the test
        - name: since
          in: query
          type: string
          description: >-
            The start date of when measurements were run (ex.
            "2016-10-20T10:30:00")
        - name: until
          in: query
          type: string
          description: >-
            The end date of when measurement were run (ex.
            "2016-10-20T10:30:00")
        - name: since_index
          in: query
          type: string
          description: Return results only strictly greater than the provided index
        - name: order_by
          in: query
          type: string
          description: |-
            by which key the results should be ordered by (default: `test_start_time`)
          enum:
            - test_start_time
            - probe_cc
            - report_id
            - probe_asn
            - test_name
            # These are all equivalent
            - index
            - idx
            - report_no
        - name: order
          in: query
          type: string
          description: |-
            If the order should be ascending or descending (one of: `asc` or `desc`)
          enum:
            - asc
            - desc
            - ASC
            - DESC
        - name: offset
          in: query
          type: integer
          description: 'Offset into the result set (default: 0)'
        - name: limit
          in: query
          type: integer
          description: 'Number of records to return (default: 100)'
      responses:
        '200':
          description: List of files that match the requested criterial with pagination
          schema:
            $ref: "#/definitions/FileList"

        default:
          description: Default response

  /version:
    get:
      summary: ''
      description: ''
      operationId: 'measurements.api.get_version'
      responses:
        '200':
          description: Returns version information
          schema:
            $ref: "#/definitions/VersionInfo"


        default:
          description: Default response

definitions:
  VersionInfo:
    properties:
      version:
        type: string
        description: The current version of the measurements API

    example:
      version: '1.0.0'

  PaginationMetadata:
    properties:
      offset:
        type: integer
        description: an integer specifying the current offset into the data
      limit:
        type: integer
        description: an integer specifying how many results should be presented
      count:
        type: integer
        description: an integer expressing the total number of items

      pages:
        type: integer
        description: >
            The number of pages, or the number of requests you will
            have to do with the current value of limit to obtain the
            full set of records
      next_url:
        type: string
        description: the url to be used to fetch the next set of items

  FileList:
    properties:
      metadata:
        $ref: "#/definitions/PaginationMetadata"
      results:
        $ref: "#/definitions/File"
    example:
      metadata:
        offset: 0
        limit: 100
        count: 420
        pages: 5
        next_url: 'https://api.ooni.io/api/v1/files?offset=100&limit=100'
      results:
        - download_url: 'https://s3.amazonaws.com/ooni-public/sanitised/2017-09-04/20170904T021115Z-US-AS7922-ndt-20170903T164039Z_AS7922_tx4Us3Eddf4XZYu9eFBiJrWq6FyS5dMLuAhF2lDx6AZs2XpGz1-0.2.0-probe.json'
          index: 754699
          probe_asn: AS7922
          probe_cc: US
          test_start_time: '2017-09-04T02:11:15Z'

  File:
    properties:
      probe_asn:
        $ref: "#/definitions/ProbeASN"
      probe_cc:
        $ref: "#/definitions/ProbeCC"
      test_name:
        $ref: "#/definitions/TestName"
      index:
        type: integer
        description: the index of this result (useful when using since_index)
      test_start_time:
        $ref: "#/definitions/TestStartTime"

      download_url:
        type: string
        description: >-
          Url to the download. Note: if the download URL ends
          with `.gz` it should be considered compressed with gzip.

  MeasurementList:
    properties:
      metadata:
        $ref: "#/definitions/PaginationMetadata"
      results:
        $ref: "#/definitions/MeasurementMetadata"

    example:
      metadata:
        offset: 0
        limit: 100
        count: 420
        pages: 5
        next_url: 'https://api.ooni.io/api/v1/measurements?offset=100&limit=100'
      results:
        - anomaly: true
          confirmed: false
          failure: false
          measurement_url: 'https://api.ooni.io/api/v1/measurement/xxxx-yyy-zzz-dddd-eeee-ffff'
          measurement_id: 'xxxx-yyy-zzz-dddd-eeee-ffff'
          measurement_start_time: '2017-08-23T14:22:13Z'
          input: 'https://ooni.io/'
          report_id: '20170823-XXX-YYY'
          test_name: 'web_connectivity'
          probe_cc: 'IT'
          probe_asn: 'AS0'

  MeasurementMetadata:
    properties:
      measurement_id:
        type: string
        description: The ID of the measurement

      measurement_url:
        type: string
        description: The URL used to retrieve the measurement blob

      measurement_start_time:
        $ref: "#/definitions/MeasurementStartTime"

      test_name:
        $ref: "#/definitions/TestName"

      input:
        $ref: "#/definitions/Input"

      probe_cc:
        $ref: "#/definitions/ProbeCC"

      probe_asn:
        $ref: "#/definitions/ProbeASN"

      confirmed:
        $ref: "#/definitions/ConfirmedAnomaly"

      anomaly:
        $ref: "#/definitions/Anomaly"

      failure:
        $ref: "#/definitions/MeasurementFailure"

  ConfirmedAnomaly:
    type: boolean
    description: |-
      Will be true for confirmed network anomalies (we found a blockpage, a middlebox was found, the IM app is blocked, etc.)

  Anomaly:
    type: boolean
    description: |-
        Measurements that require special attention (it's likely to be a case of blocking), however it has not nessarily been confirmed

  MeasurementFailure:
    type: boolean
    description: |-
        There was an error in the measurement (the control request failed, there was a bug, etc.)

  TestStartTime:
    type: string
    description: |-
      Timestamp of when the test was started (many measurements) in
      UTC time coordinates (ex. 2015-08-24 12:02:23) (Note: ooniprobe <=
      1.4.0 generates skewed timing information)

  MeasurementStartTime:
    type: string
    description: |-
      Timestamp of when the measurement was performed in
      UTC time coordinates (ex. `2015-08-24 12:02:23`) (Note: ooniprobe <=
      1.4.0 generates skewed timing information).
      This timestamp is more accurate than the `test_start_time` field

  Input:
    type: string
    description: |-
      If the test takes an input this will contain the value of it
      these can be for example URLs, hostnames, IPs, etc.

  ProbeCC:
    type: string
    description: The two letter country code of the probe or ZZ if inlcudecountry is set to false

  ProbeASN:
    type: string
    description: |-
      The AS Number of the probe (prefixed by AS, ex. AS1234)
      or AS0 if includeasn is set to false

  TestName:
    type: string
    description: The name of test that generated this measurement

  MeasurementBlob:
    properties:
      test_start_time:
        $ref: "#/definitions/TestStartTime"

      measurement_start_time:
        $ref: "#/definitions/MeasurementStartTime"

      test_keys:
        type: object

      input_hashes:
        description: This is a deprecated field
        type: array
        items:
          type: string

      probe_ip:
        type: string
        description: |-
          The IPv4 address of the probe or 127.0.0.1 if includeip is set to false

      id:
        type: string
        description: This is an identifier of this particular measurement

      test_helpers:
        type: object
        description: This is an identifier of this particular measurement

      probe_cc:
        $ref: "#/definitions/ProbeCC"

      test_runtime:
        type: number
        description: The time the test took to run

      input:
        type: string
        description: |-
          If the test takes an input this will contain the value of it
          these can be for example URLs, hostnames, IPs, etc.

      probe_city:
        type: string
        description: |-
          The name of the city of the probe or null if
          includecity is set to false (note: recent version of ooniprobe always set this to null)

      probe_asn:
        $ref: "#/definitions/ProbeASN"

      annotations:
        type: object
        description: |-
          Extra metadata (for example the platform) associated to the measurement

      software_name:
        type: string
        description: |-
          The name of the software that has generated
          such measurement (ex. ooniprobe)

      software_version:
        type: string
        description: The version of the software used to generate this report

      data_format_version:
        type: string
        description: |-
          The data format version for this measurements (currently: 0.2.0)

      report_filename:
        type: string
        description: The full filepath to the report_filename

      test_version:
        type: string
        description: The version of this test

      bucket_date:
        type: string
        description: The date in which the measurement was processed

      test_name:
        $ref: "#/definitions/TestStartTime"

      report_id:
        type: string
        description: A string that identifies the report this measurement is part of

      backend_version:
        type: string
        description: The version of the backend that collected this measurement

      options:
        type: object
        description: The version of the backend that collected this measurement